requirements.txt
# PDF Processing
PyMuPDF==1.26.0
pypdf==5.6.0
nltk==3.9.1
langchain-text-splitters==0.3.8

# ML Models
torch>=2.6.0
transformers>=4.48.0,<5.0.0
FlagEmbedding>=1.3.5

# LLM Providers
openai==1.85.0
anthropic

# Utilities
numpy>=2.0.0
pandas
python-dotenv
tqdm
